+++ src/main/java/com/academy/mudogroupware/lecture/application/usecase/UpdateLectureUseCase.java
@@ -1,0 +1,8 @@
+package com.academy.mudogroupware.lecture.application.usecase;
+
+import com.academy.mudogroupware.lecture.application.command.UpdateLectureCommand;
+
+public interface UpdateLectureUseCase {
+
+    void updateLecture(UpdateLectureCommand command);
+}